feat(add): check default branch
First attempt to implement check of default branch.

To get the setup default branch of a repository you need to query the GitHup API, which has some input parameters:

- Owner (aka GH Org)
- Repository name (e.g. tractusx-quality-checks)

The check needs to know these. When running as a GH Action/Workflow there are environment vars available, which makes it easy to retrieve these information. When running locally, it's a bit harder to extract these informations. In this first draft I decided to parse the `.git/config` file and use the `url` key.

For determining the repository name I've created the func `getRepoName()` which returns owner, repo(name) and a bool called local, which is true if `.git/config` was parsed.

A second func `getRepoInfo()` queries the GitHup API and returns `*github.Repository` object.

Finally, the func `QualityGuideline.Test()` first checks if it's executed on a fork and if true it returns the check as failed, due to unreliable circumstances. Then it tests for DefaultBranch == main and returns appropriate.